+++ bin/workflow_glue/downsampling_ratio.py
@@ -4,6 +4,7 @@
 import sys
 
 from ezcharts.components.mosdepth import load_mosdepth_summary  # noqa: ABS101
+import pandas as pd
 
 from .util import wf_parser  # noqa: ABS101
 
@@ -16,6 +17,11 @@ def argparser():
         '--summary',
         help="Mosdepth summary file",
         required=True
+    )
+    parser.add_argument(
+        '--bed',
+        help="Mosdepth region bed file",
+        required=False
     )
     parser.add_argument(
         '--downsample_depth',
@@ -35,8 +41,20 @@ def argparser():
 
 def main(args):
     """Run the entry point."""
-    depth_t, depth_r = load_mosdepth_summary(args.summary)
-    value = depth_t.loc[depth_t['chrom'] == 'total', 'mean'].values[0]
+    # CW-2799: If a region bed is provided, compute the average coverage in the
+    # regions only. We leverage the length of each interval by its coverage to reduce
+    # the weights of the shorter regions vs the longer.
+    if args.bed:
+        df = pd.read_csv(args.bed, names=['chrom', 'start', 'end', 'depth'], sep="\t")
+        df.eval('length=end-start', inplace=True)
+        df.eval('coeff=length*depth', inplace=True)
+        value = df.coeff.sum() / df.length.sum()
+    # If no bed is provided, simply use the mosdepth summary as a proxy. Use the
+    # total_region rather than the total, even though in this case they should be
+    # the same.
+    else:
+        depth_t, depth_r = load_mosdepth_summary(args.summary)
+        value = depth_r.loc[depth_r['chrom'] == 'total_region', 'mean'].values[0]
     # Check if the value is at least 10% greater than the expected depth.
     # If it is, compute the downsample ratio. If not, return -1.
     margin = args.margin if args.margin >= 1.0 else 1.0